+++ src/apps/sablier/ethereum/sablier.stream.contract-position-fetcher.ts
@@ -143,7 +143,7 @@ export class EthereumSablierStreamContractPositionFetcher extends CustomContract
           },
 
           displayProps: {
-            label: `${isRecipient ? 'Available' : 'Deposited'} ${token.symbol} on Sablier`,
+            label: `${isRecipient ? 'Available' : 'Deposited'} ${token.symbol} on Sablier stream #${stream.streamId}`,
             secondaryLabel: buildDollarDisplayItem(token.price),
             images: getImagesFromToken(token),
           },