+++ src/apps/teahouse/arbitrum/teahouse.vaults.token-fetcher.ts
@@ -3,12 +3,12 @@ import { PositionTemplate } from '~app-toolkit/decorators/position-template.deco
 import { TeahouseVaultsTokenFetcher } from '../common/teahouse.vaults.token-fetcher';
 
 @PositionTemplate()
-export class OptimismTeahouseVaultsTokenFetcher extends TeahouseVaultsTokenFetcher {
+export class ArbitrumTeahouseVaultsTokenFetcher extends TeahouseVaultsTokenFetcher {
   groupLabel = 'Vault share';
 
-  queryFilterFromBlock = 53274656;
+  queryFilterFromBlock = 84086498;
 
   async getAddresses() {
-    return ['0x9ae039f9de94542f6f1b3fba60223e6aa4f411af', '0xee1e02609a480bdc9d9651c200d90222b6691f03'];
+    return ['0x9f4fff022ebff0cbfa3faf702911d0f658a4ba9b'];
   }
 }